refactor: 홈 화면 이벤트 핸들러 분리
